Move Profile and Action structure under Models directory.
