+++ source/QuickCut/CMakeLists.txt
@@ -51,8 +51,10 @@ set(CXX_FILES
     ${MISC_CXX_FILES}
     ${UI_CXX_FILES}
     ${WIDGETS_CXX_FILES}
-    ${QTSERVICE_CXX_FILES}
+    # Shared sources:
     ${SHARED_CXX_FILES}
+    ${QUICKCUT_MODELS_CXX_FILES}
+    ${QTSERVICE_CXX_FILES}
 )
 
 set(QRC_FILES
@@ -118,6 +120,9 @@ source_group("Source Files\\Misc" FILES
 )
 source_group("Source Files\\Shared" FILES
     ${SHARED_CXX_FILES}
+)
+source_group("Source Files\\Shared\\Models" FILES
+    ${QUICKCUT_MODELS_CXX_FILES}
 )
 source_group("Source Files\\Shared\\QtService" FILES
     ${QTSERVICE_CXX_FILES}
