+++ source/QuickCutConsole/CMakeLists.txt
@@ -22,10 +22,6 @@ include_directories(
 ##################################################################################################################
 set(SHARED_CXX_FILES
     ${QUICKCUT_SHARED_DIR}/pch.h
-    ${QUICKCUT_SHARED_DIR}/Profile.h
-    ${QUICKCUT_SHARED_DIR}/Profile.cpp
-    ${QUICKCUT_SHARED_DIR}/Action.h
-    ${QUICKCUT_SHARED_DIR}/Action.cpp
     ${QTSINGLEAPP_CXX_FILES}
 )
 set(MISC_CXX_FILES
@@ -57,6 +53,7 @@ endif()
 set(CXX_FILES
     ${MISC_CXX_FILES}
     ${SHARED_CXX_FILES}
+    ${QUICKCUT_MODELS_CXX_FILES}
 )
 
 ##################################################################################################################
@@ -67,6 +64,9 @@ source_group("Source Files" FILES
 )
 source_group("Source Files\\Shared" FILES
     ${SHARED_CXX_FILES}
+)
+source_group("Source Files\\Shared\\Models" FILES
+    ${QUICKCUT_MODELS_CXX_FILES}
 )
 source_group("Resources" FILES
     ${RESOURCE_FILES}
