+++ source/QuickCutConsole/QuickCutConsoleWindows.cpp
@@ -2,7 +2,7 @@
 
 #include "QuickCutConsoleWindows.h"
 #include "pch.h"
-#include "Profile.h"
+#include "Models/Profile.h"
 
 #include <QDebug>
 #include <QDir>
