+++ source/QuickCutShared/Models/Action.cpp
