+++ source/QuickCutShared/Models/Action.h
