+++ source/QuickCutShared/Models/CMakeLists.txt
@@ -1,0 +1,13 @@
+
+cmake_minimum_required(VERSION 3.12)
+
+set(QUICKCUT_MODELS_DIR ${CMAKE_CURRENT_SOURCE_DIR} CACHE STRING "")
+
+file(GLOB QUICKCUT_MODELS_CXX_FILES
+    *.h
+    *.cpp
+)
+set(QUICKCUT_MODELS_CXX_FILES
+    ${QUICKCUT_MODELS_CXX_FILES}
+    CACHE STRING "" FORCE
+)
