+++ source/QuickCutShared/Models/Profile.cpp
