+++ source/QuickCutShared/Models/Profile.h
